First commit.
A simple timer. The main inspiration was for use as a Pomodoro timer,
but I can see it being widely applicable to other domains.

Right now, the functionality is very limited. The alarm can't be run
in the background. Also, there is no interactivity for convenient
usage for a full Pomodoro session of four bouts of study.